Script/requirements.txt
pbkdf2
dpapick
M2Crypto==0.22.3
ssdeep
python-registry
